+++ src/Symfony/Component/Security/Core/CHANGELOG.md
@@ -25,6 +25,7 @@ CHANGELOG
 7.4
 ---
 
+ * Add `StatelessTokenInterface` to mark tokens that should not be persisted to the session
  * Add `MermaidDumper` to dump Role Hierarchy graphs in the Mermaid.js flowchart format
  * Deprecate `PersistentTokenInterface::getClass()`, the user class will be removed from the remember-me cookie in 8.0
  * Add `extraData` property to `Vote` objects